+++ UI/ChallengeDetails/Tabs/Mechanics/Tabs/Deaths.lua
@@ -44,20 +44,23 @@ function MyDungeonsBook:DeathsFrame_Create(parentFrame, challengeId)
     deathLogTable:RegisterEvents({
         OnEnter = function (_, cellFrame, data, _, _, realrow, column)
             if (realrow) then
-                if (column == 5) then
-                    self:DeathsFrame_LogCellHover(cellFrame, data[realrow].cols[5].value);
+                if (column == 1) then
+                    self:DeathsFrame_ArgoCellHover(cellFrame, data[realrow].cols[column].value);
                 end
                 if (column == 6) then
+                    self:DeathsFrame_LogCellHover(cellFrame, data[realrow].cols[column].value);
+                end
+                if (column == 7) then
                     self:DeathsFrame_AmountCellHover(cellFrame, deathsFrame:GetUserData("deathLogTableRowData")[realrow]);
                 end
-                if (column == 7 or column == 8) then
-                    self:Table_Cell_SpellMouseHover(cellFrame, data[realrow].cols[1].value);
+                if (column == 8 or column == 9) then
+                    self:Table_Cell_SpellMouseHover(cellFrame, data[realrow].cols[column].value);
                 end
             end
         end,
         OnLeave = function (_, _, _, _, _, realrow, column)
             if (realrow) then
-                if (column >= 5 and column <= 8) then
+                if (column == 1 or column >= 6 and column <= 9) then
                     self:Table_Cell_MouseOut();
                 end
             end
@@ -227,8 +230,10 @@ function MyDungeonsBook:DeathsFrame_GetDataForDeathLogTable(challengeId, key, un
         local source = getLogItemByIndex(log, 5);
         local subEventPrefix, subEventSuffix = subEventName:match("^(.-)_?([^_]*)$");
         local hp = "";
+        local threat = -1;
         if (type (log[1]) == "table") then
             hp = log[1].hp.current .. "=" .. log[1].hp.max;
+            threat = log[1].agro and log[1].agro.status or -1;
         end
         if (subEventName:match("SPELL_AURA_REMOVED")) then
             amount = "-" .. getLogItemByIndex(log, 15); -- BUFF or DEBUFF
@@ -259,6 +264,7 @@ function MyDungeonsBook:DeathsFrame_GetDataForDeathLogTable(challengeId, key, un
         if (subEventName == "SPELL_DISPEL") then
             spellId = getLogItemByIndex(log, 12);
         end
+        tinsert(row.cols, {value = threat});
         tinsert(row.cols, {value = spellId});
         tinsert(row.cols, {value = unitId});
         tinsert(row.cols, {value = tonumber(string.format("%.2f", getLogItemByIndex(log, 1) - lastLogTime))});
@@ -279,10 +285,18 @@ function MyDungeonsBook:DeathsFrame_GetColumnsForDeathLogTable(challengeId)
     return {
         {
             name = "",
+            width = 20,
+            align = "RIGHT",
+            DoCellUpdate = function(...)
+                self:Table_Cell_FormatCellAsArgo(...);
+            end
+        },
+        {
+            name = "",
             width = 25,
             align = "LEFT",
             DoCellUpdate = function(...)
-                self:Table_Cell_FormatCellAsDeathLogAvoidable(challengeId, ...)
+                self:Table_Cell_FormatCellAsDeathLogAvoidable(challengeId, ...);
             end
         },
         {
@@ -405,7 +419,7 @@ end
 ]]
 function MyDungeonsBook:Table_Cell_FormatCellAsDeathLogAvoidable(challengeId, _, cellFrame, data, _, _, realrow, column)
     local spellId = data[realrow].cols[column].value;
-    local unitId = data[realrow].cols[2].value;
+    local unitId = data[realrow].cols[3].value;
     local isAvoidable = self:IsSpellAvoidableForPartyMember(challengeId, unitId, spellId) and (data[realrow].cols[6].value ~= "-DEBUFF");
     local icon = "interface\\raidframe\\readycheck-notready.blp";
     local suffix = self:GetIconTextureSuffix(16);
@@ -418,7 +432,7 @@ end
 function MyDungeonsBook:DeathsFrame_AmountCellHover(cellFrame, log)
     local logMsg, r, g, b = CombatLog_OnEvent(DEFAULT_COMBATLOG_FILTER_TEMPLATE, unpack(log, (type(log[1]) == "number" and 1) or 2));
     GameTooltip:SetOwner(cellFrame, "ANCHOR_NONE");
-    GameTooltip:SetPoint("TOPLEFT", cellFrame, "BOTTOMLEFT");
+    GameTooltip:SetPoint("TOPLEFT", cellFrame, "BOTTOMRIGHT");
     GameTooltip:SetText(logMsg or "", {r = r, g = g, b = b});
     GameTooltip:Show();
 end
@@ -428,8 +442,18 @@ end
 ]]
 function MyDungeonsBook:DeathsFrame_LogCellHover(cellFrame, subEventName)
     GameTooltip:SetOwner(cellFrame, "ANCHOR_NONE");
-    GameTooltip:SetPoint("TOPLEFT", cellFrame, "BOTTOMLEFT");
+    GameTooltip:SetPoint("TOPLEFT", cellFrame, "BOTTOMRIGHT");
     GameTooltip:SetText(subEventName);
+    GameTooltip:Show();
+end
+
+--[[--
+@local
+]]
+function MyDungeonsBook:DeathsFrame_ArgoCellHover(cellFrame, argoLevel)
+    GameTooltip:SetOwner(cellFrame, "ANCHOR_NONE");
+    GameTooltip:SetPoint("TOPLEFT", cellFrame, "BOTTOMRIGHT");
+    GameTooltip:SetText(self:GetArgoStatus(argoLevel));
     GameTooltip:Show();
 end